add support for streaming svc-js handlers
flow:

when a request arrives, we check if we have a stream handler for
it, given the new `.stream()` method sent the action name and
request id via headers. if a handler with the suffix "Stream"
exists, we call the stream handler. If not, we buffer the stream
and call a regular action handler.